requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
environs==9.5.0
marshmallow==3.20.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
